feat: add login feature and a lot more (read long message)
so basically, I had to create a way to manage certain global values (token and user profile data). This was done by using Cubits. I also created an http interceptor and changed all services to work based on it, to sort of standarize the way we create services and prevent hard coding the reading of the Bearer token. Lastly, I coded the login feature. To do so, I had to instally the supabase dependency and ensure it initializes on the main app so the Auth service (which uses the supabase client) works as intended and we can both retrieve the authorization token and keep it updated even after closing the app safely. This is probably the largest commit I have ever done. May christ be with us.