chore: removed unused favorites page (read long message)
the favorites feature was planned to be a view nested inside the 'you' feature. As such, it doesnt deserve its own folder. To add salt on the wound this wasnt even implemented on a real view widget, and I find no point on using it as a base to create one since I have the HomePage which already implements a lot of the logic and view I need. I also refactored the product card into the shared folder in anticipation to the incoming change